refactor: Maya環境に依存しないベクトル演算関数をutils.pyに移動
- cross_product, normalize_vector, pmx_to_maya_vectorをmaya_utilsからutilsへ移動
- Maya環境に依存しない汎用的な関数をutilsモジュールに集約
- rig_converterでの参照を更新
- CLAUDE.mdにユーティリティモジュールの使用方法を追記
